Create Profile one page, no phone number, and other stuff
Create Profile is now one page, no phone number across the app, edit profile is now preloaded, other minor tweaks
